fixes, teleport usage in city
- vehicles can use teleport when moving by holding ctrl (teleport is
inaccurate so it doesn't become OP)
- usage of teleporter when vehicle retreats is planned for
- list control no longer passes clicks to items that are out of bounds
and not displayed
- vehicle selection box now allows scrolling for more vehicles

+++ framework/render/gles30_v2/ogles_3_0_renderer_v2.cpp
@@ -309,7 +309,7 @@ class Spritesheet
 		while (!validEntries.empty())
 		{
 			LogInfo("Repack: creating sheet %d", (int)pages.size());
-			auto page = mksp<SpritesheetPage>(pages.size(), page_size, node_count);
+			auto page = mksp<SpritesheetPage>((int)pages.size(), page_size, node_count);
 			pages.push_back(page);
 			page->addMultiple(validEntries);
